requirements.txt
requests == 2.21.0
boto3 == 1.9.135
paramiko == 2.4.2
psycopg2-binary == 2.8.2
colored == 1.3.93
